bug fixes and code improvement
Removed some unnecessary else if conditions and added a few throw statements to improve the detection of potential new errors.
Regained GUI button and checkbox logic (elements that conflict each other are disabled when the right conditions are met).
Fixed a few bugs that had to do with the format selection (still WIP !).