fixed a ton of bugs with the setup
The script will now use a workaround to install yt-dlp and create the necessarry environmental variables in the system.
A bug has been fixed where the script does not contain a recognized action when the user or an error cancels any of the open command prompt windows.
The script will now check if a user has an established Internet connection and react upon the result.
Several other code improvements.
BUGS ARE TO BE EXPECTED (WIP!)