fixed the subtitle download option
The script will now tell yt-dlp to actually download all available subtitles.
With this function now working a new error occured which caused the progress bar to fill up to fast (again).
This has been fixed with a lock to the progress bar until the actual video file download is detected via a message.