HUGE rework for .MSI installation
Removed all files that have been stored into the main .EXE file.
Those files are now included into the .MSI installer.
FFmpeg is now downloaded using a PowerShell script.
The setup which cannot be done via the .MSI file has been outsourced into a seperate Setup.exe file which controls the setup action. It can be called via any command line with a few parameters specified into a text file right next to it.
I also made some changes to the overall script for example some MsgBox titlles or symbols have been unified to create a better look, but I expect to find more small things like that in the future.

I testet the whole installer in a virtual machine and it works just fine. The installation is far from clean but it works for now. Maybe I will add an extra GUI for that who knows.
Next step is going to be the refinement of the setup and the AdvancedInstaller integration for my GitHub workflows. Yessir