further progress with .MSI installation
The setup works just fine on a VM assuming it will run on allmost any Windows 7 or 10 machine. It will also detect if all requirements are present in case the software has been installed before. The amount of prompts has been minimized to make the setup as simple as possible.