added download folder cleanup option
When a download does not produce any files into the media folder the script will now delete the folder which contains the media folder and other files. This option can be toggled via the config file.
The workflow has been adjusted so that the pre releases are now named differently.