first successful installation with PSDeployToolkit
From now on PowerShell AppDeployToolkit will be used to install VideoDownloader and it's dependencies.
I recoded the initial PowerShell setup scripts to work with exit codes and more parameters. I still use a .MSI file for the main application because it is very comfortable and easy to change.
I will remove the VideoDownloaderSetup.exe in the next few commits and integrate the whole logic into the PS Toolkit.
Eventough the installation process works fine, I expect a lot of hidden bugs / copy & paste issues to reveal from time to time :D