further installation process improvements
The setup should work fine now most of the time.
The integrity checks have been improved when launching the script.
It is now possible to repair the installation from the main GUI.
The CI workflow has been updated to generate a compressed archive for installation purposes.
A few refactoring changes will be made in the future on a different branch.