fixed repair action and small code issues
After successfully repairing the script with the PowerShell setup, the registry variables booleanFirstTimeLaunch and booleanSetupRequired were not set to the correct values, which forced the user to repair the script again, even though the repair was completed with no errors.
Furthermore some small adjustment have been made for example moving all file related methods to the FileManager.AHK file.