first successful install
The program can now be installed purely using a .MSI installer file instead of the PS App Deploy Toolkit.

The installation is not as clean as I hoped it would be. There are to many message boxes at the first launch.
The automatic update check is missing as well.

There WILL be bugs and errors.