first implementation of video list GUI
The video list GUI contains a list view element which stores the selected videos with their corresponding thumbnail and title as well as their duration and uploader.
It also contains a search bar to search the videos in the list. Most of the features are WIP, which means they won't work or are disabled partially.