ported URL capture hotkeys
The URL capture hotkeys will now add the captured video URL to the video list instead of writing them to the URL file (which will be removed later on).
I also made a function to validate video URLs and cleaned up the code a bit.
The hotkey which extracts the URL from the video under the mouse cursor is still very unreliable. I might have to change the whole mechanism to fix this.