improved automatic update backup
When updating the script via the prompt, the old version will be backed up into a separate folder.
The backup will now exclude all temporary directories and the default download folder.
Files larger than 10MB will be excluded as well to keep the backup as small as possible.

I also fixed an issue where a few hotkeys for sub menus in the video list GUI made it difficult to enter numbers.