updated config file handler
I improved the way that the program deals with it's config file.
The script will now recognize a damaged or old config file from a different version of VideoDownloader and create a new one. The catch is that it will try to import as many of the settings from the old file as possible to avoid data loss.

The idea of separating the main and working directory has been abandoned because the working directory was only ever used to provide a root for the download and temp folders which are now implemented in the new version of the default config file.