fixed tutorial control highlight issue
When a tutorial highlighted a control or menu element, the border lines would appear to long. This was caused by the OS trying to scale each border (which is technically a GUI) by the set percentage.

For example:

A screen has set it's scale to 150%. The menu element would be scaled by the OS automatically by the factor 1.5.
The interactive tutorial would gather the dimensions of the already scaled element and the OS would scale the 4 GUIs again, causing the borders to become longer than the menu element.