improved graphics and button logic
A few buttons have been colored to make them stand out. This is possible by using the "ColorButton" library made by @nperovic. Thanks for that!
Some buttons are disabled now when they do not make sense to press as well.

The settings GUI has been changed a bit and I changed the default download directory to the user's download folder.