improved config file
There was a bug causing the application to mistakenly display the notification that an old or corrupted config file has been imported when installing VideoDownloader for the first time.
I also fixed an issue which caused the config file sections to always be in a random order when a default config file was created.