added additional exit functionality
VideoDownloader will now terminate entirely when the video list GUI is closed. If there are any videos in the video list, the user will be prompted to confirm the termination.
There is an option to toggle this behavior in the settings GUI which makes it, so that the application continues to run in the background when the video list GUI is closed.