improved exit functionality
When VideoDownloader is closed, it will now warn the user about an ongoing download process as well.
The sub menu [Exit Application] of the [Actions] menu now closes VideoDownloader with warning messages instead of forcing the termination.
The termination hotkey is currently the only built-in option to close VideoDownloader without any prompts.