improved GUI positioning
I fixed a bug which caused the video list GUI to flicker whenever it was maximized when the last state was loaded.
The tutorial GUIs will no be positioned to interfere less with the relevant elements.
The settings and help GUI should now position themselves above the video list window, even when it is on a different monitor.