bug fixes and improvements
The config file has not been initialized when the setup called the checkInternetConnection() function which caused an error. I changed the way that the application reads the value from the config file in this case and moved the declaration of the config file location as a result.
The image of the video list GUI has been updated as well because resizing is now supported, which changes the look of the GUI a bit.