improved file versioning
The executable file generated by the AutoHotkey compiler will now have the release version in its file attributes.
The application no longer extracts its version from the registry but from the VideoDownloader executable file directly (if possible).

It should be possible to download the source code and execute the .AHK file directly or to compile the VideoDownloader.ahk file and run the executable by now.