improved dependency management
It is now easier to select custom yt-dlp and FFmpeg executables. A file selection prompt will be displayed which then copies the selected executables into the specific VideoDownloader sub folder.
I added a menu element in the [Actions] menu to open the setup window as well as to update all dependency executables.