improved video list search
The videos in the video list now store the original URL used to find them.
In other words, if you add a video with a playlist URL, you would be able to find the video either by it's original URL on YouTube for example or by searching for the playlist URL.
The code has been cleaned up as well.