improved multiple video options
Changing the desired format or subtitle while multiple videos are selected causes a dialog box to appear. If confirmed, all changes made to the currently selected video will be applied to the other selected videos in the video list.

Holding the SHIFT while closing the video list GUI saves it's position and state regardless whether the application is configured to exit or not.

A few code and performance improvements have been made as well.