fixed double config file issue
When the program was launched for the very first time, there is a chance that not all dependencies are present. If that is the case, the user is forced to either provide or download them, which causes VideoDownloader to restart.
In this case it did not exit right away because there was enough time to create an empty version of the config file which lead to the "import message" even if the application was seemingly launched for the first time.
I fixed a bug which did not allow to download the source code and compile the "VideoDownloader.ahk" file afterwards to an executable without issues.